input_study.py float으로 변경

+++ chpater06/input_study.py
@@ -1,3 +1,3 @@
 b = input("정사각형 한 변의 길이는?:")
-area = int(b) ** 2
+area = float(b) ** 2
 print("정사각형의 넓이는 : {}".format(area))
